Adding search message, minor debugging and design pump
Adding a search message whenever fulltext search and timesearch is in progress,
i.e. whenever the .../full/... API endpoint is being rewuested.
The message also stays durning client-side rendering (not only during the ajax call).

During bugfix it was noticed, that in order to exit a jquery each loop,
one has to "return false". A mere "return" will just skip to the next iteration.
This led to the bug where the currently playing song wouldn't be shown when the
reset button was pressed.

Also desided to pump client-side release versions separately from server-side release versions
since those are two different products and are developed indipendently. It is important to not though,
that the Dockerfile, which is stored in the server-side code repository, needs to be adjusted whenever
a new client is released.

This is a v1.2.1-pre-release